requirements.txt
biopython==1.83; python_version >= "3.8"
certifi==2024.2.2; python_version >= "3.7"
charset-normalizer==3.3.2; python_full_version >= "3.7.0" and python_version >= "3.7"
idna==3.7; python_version >= "3.7"
numpy==1.23.0
pandas==1.4.3; python_version >= "3.8"
python-dateutil==2.8.2; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
pytz==2022.1; python_version >= "3.8"
requests==2.31.0; python_version >= "3.7"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
textwrap3==0.9.2
urllib3==2.2.1; python_version >= "3.8"
